+++ app/Http/Requests/Country/UpdateCountryRequest.php
@@ -2,7 +2,6 @@
 
 namespace App\Http\Requests\Country;
 
-use App\Rules\Country\CodeHasProperLengthRule;
 use App\Rules\Country\CountryTitleHasProperLengthRule;
 use App\Rules\Film\FilmExistsRule;
 use App\Rules\User\UserExistsRule;
@@ -12,7 +11,6 @@ use Illuminate\Foundation\Http\FormRequest;
  * Class UpdateCountryRequest
  * @package App\Http\Requests\Country
  * @property string $title
- * @property string $code
  * @property string[] $films
  * @property string[] $users
  */
@@ -27,7 +25,6 @@ class UpdateCountryRequest extends FormRequest
     {
         return [
             'title' => ['bail', 'required', 'string', new CountryTitleHasProperLengthRule()],
-            'code' => ['required', 'string', new CodeHasProperLengthRule()],
             'films' => ['nullable', 'array'],
             'films.*' => ['bail', 'required', 'integer', new FilmExistsRule()],
             'users' => ['nullable', 'array'],