+++ database/seeds/ProfessionSeeder.php
@@ -15,26 +15,32 @@ class ProfessionSeeder extends Seeder
     {
         $professions = [
             [
+                'id' => 0,
                 'title' => 'Актер',
                 'plural_title' => 'Актеры'
             ],
             [
+                'id' => 1,
                 'title' => 'Композитор',
                 'plural_title' => 'Композиторы'
             ],
             [
+                'id' => 2,
                 'title' => 'Оператор-постановщик',
                 'plural_title' => 'Операторы-постановщики'
             ],
             [
+                'id' => 3,
                 'title' => 'Продюсер',
                 'plural_title' => 'Продюсеры'
             ],
             [
+                'id' => 4,
                 'title' => 'Режиссер',
                 'plural_title' => 'Режиссеры'
             ],
             [
+                'id' => 5,
                 'title' => 'Сценарист',
                 'plural_title' => 'Сценаристы'
             ]
@@ -42,6 +48,7 @@ class ProfessionSeeder extends Seeder
 
         foreach ($professions as $profession) {
             Profession::create([
+                'id' => $profession['id'],
                 'title' => $profession['title'],
                 'plural_title' => $profession['plural_title']
             ]);